Convert updates to protocol/message data types (#253)
* Convert updates to protocol/message data types

* Reconsidering the definition of Outcome to exclude Incomplete

The possibility for an outcome to be incomplete is not present in the
protocol messages - it is only something that can happen with the RPC so
we move that indication up a level to the RPC response object. This also
prevents us from accidentally storing Incomplete{} as the outcome value
in an update completed event.

* Add workflow task failed causes

* Add ProtocolMessageCommand

Will be used to reference and sequence protocol messages from the
RespondWorkflowTaskCompletedRequest.Messages field.