Add workflow expiration and run timestamps to execution info (#505)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
Add new structure -WorkflowExecutionExtendedInfo

That structure has few fields:
* execution timeout time
* run timeout time
* last reset time
* cancel requested

(the last one from https://github.com/temporalio/api/pull/339)

<!-- Tell your future self why have you made these changes -->
**Why?**
Per customer request.
Execution timeout timestamp may change after workflow reset, and in this
case customers have no idea when it will fire.
WorkflowExecutionInfo is reused in other server calls, like
ListWorkflow, etc.
Because of that we either add any new field to ES, or it will be nil
which is bad user experience.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
No

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**
N/A