[Requires repository secret addition] Add Publish to Buf Github Action (#508)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
Adds a GitHub Actions workflow to automate the process of pushing
protocol buffers to the Buf registry on every push to the main branch.

We need to set the `BUF_TOKEN` secret before merging this PR.

<!-- Tell your future self why have you made these changes -->
**Why?**
We want to be able to depend on server protos from other repositories
(e.g. api-cloud) using the buf package manager.


<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
None.

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**
NA