New Deployment protos to work with already-implemented versioning 3 API's (#515)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
- Updated protos to reflect changes based on the recent conclusion
regarding naming conventions for versioning-3
- `Deployment` -> `WorkerDeploymentVersion`
- Added comments to remove pre-release versioning-3 code.

Note: This API does not update pre-release versioning-3 changes related
to dataplane. That currently is being worked in separate
[PR](https://github.com/temporalio/api/pull/514). The aim of this PR is
to update those protos which are required by the deployment entity
workflows so that work on those can commence there ASAP.

Additionally, the updates to the proto are not finalized and will bring
in more additions. Right now, they contain the bare minimum
fields/structs required to make the deployment entity workflows run.

With that being said, care has been taken to not add fields here and
delete them in the near future. In other words, future changes will only
bring in additions to these new protos.


<!-- Tell your future self why have you made these changes -->
**Why?**
- Versioning-3


<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**
- The Server PR is currently being built. Note, the server PR will only
contain renamings (to agree with the namings of this PR) so I don't
think that should serve as a blocker for reviewing this.

---------

Co-authored-by: Shahab Tajik <shahab@temporal.io>