Add data plane fields w/ renamed Worker Versioning concepts (#514)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
The following concepts in Worker Versioning V3 is being renamed. This PR
adds the data plane fields with new names so SDK can populate them like
the old ones.
`DeploymentSeries` -> `WorkerDeployment`
`Deployment` -> `WorkerDeploymentVersion`

Also `WorkerVersionCapabilities` and its `use_versioning` boolean are
replaced with `WorkerDeploymentOptions` and `WorkflowVersioningMode`
enum to be more compatible with V3 changes and also extensible in the
future.

`RespondActivityTask*Request` used to send `deployment` but that is now
deprecated. Since Server does not have any anticipated use for these
values, I did not add a replacement for them at the moment.

<!-- Tell your future self why have you made these changes -->
**Why?**
See ^^

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
Not breaking, some things are deprecated that will be cleaned up at a
later time with other old Versioning stuff.

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**
Server PR will follow soon. But these changes should not break server
compile because methods are not touched.

---------

Co-authored-by: Carly de Frondeville <carly.defrondeville@temporal.io>