add unset_versioning_override to WorkflowExecutionOptionsUpdatedEventAttributes (#516)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
### What:

Add unset_versioning_override to
WorkflowExecutionOptionsUpdatedEventAttributes

<!-- Tell your future self why have you made these changes -->
### Why:

So that users of this event don't need to load VersioningOverride from
mutable state every time they create this event.
This change was prompted because the event is now being used for
non-version-override-related things, and I received feedback that it is
inefficient / awkward / error-prone to have to load and pass in the
current versioning override every time anyone writes to this event.

Now, a nil Versioning Override in this event means "no change" instead
of "remove".
This reduces the chance that someone accidentally unsets an override in
the future, and also is more efficient.
We've discussed this change internally in the server team and are ok
with changing the meaning of this history event, because it is such a
small change and the scope of impact is small (pre-release versioning
users who have unset a versioning override and are building mutable
state from that history).


<!-- Are there any breaking changes on binary or code level? -->
### Breaking changes?

Now, a nil Versioning Override in this event means "no change" instead
of "remove".
If an event exists with the previous meaning and the mutable state is
rebuilt, the Versioning Override would not be removed.
But the chance of that happening is very low.

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
https://github.com/temporalio/temporal/pull/7091