some changes to ManageActivities API (#526)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
1. Remove "workflow_id" and "run_id", add "workflow_execution"
2. Combine "activity_id" and "activity_type" into a single "oneof"
message. They are mutually exclusive.

<!-- Tell your future self why have you made these changes -->
**Why?**
Feedback in previous PR.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**