Versioning 3 API refinements (#536)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
- `version` string fields in `WorkerDeploymentOptions` and
`WorkerDeploymentVersion` changed to `build_id`.
- `WorkerDeploymentVersionInfo` and `WorkerDeploymentInfo` messages
completed with all the fields that will be implemented before Replay.
- Added `conflict_token` to relevant APIs.
- Now using special `__unversioned__` value instead of empty string to
represent unversioned workers.

Note: the following APIs are still to be added:
- `UpdateWorkerDeploymentVersionMetadata`
- `DeleteWorkerDeploymentVersion`
- `DeleteWorkerDeployment`

<!-- Tell your future self why have you made these changes -->
**Why?**
Refinements based on latest design discussions within the crew.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**

---------

Co-authored-by: Carly de Frondeville <carly.defrondeville@temporal.io>