Use string values for deployment version (#547)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
- Using string type in all the places for `version` fields. Removed
`WorkerDeploymentVersion` from the public API. This is for two reasons:
- String type gives ability to specify special values of
`__unversioned__` (and maybe later `<deployment_name>/__unversioned__`.
- We might add ability for users to pass custom version IDs in with case
the struct would be hard for users to use (they'd need to write the
logic to convert the struct to an ID based on whether it has a custom ID
or not).
- Using `WorkerDeploymentOptions` in the task response APIs instead of
`WorkerDeploymentVersion` so that server has the full info regardless of
the worker participating in versioning or not.
- Added `worker_deployment` field to execution info (top level) so both
versioned and unversioned workflows can have this info (and its
corresponding SA).
- Renamed `WorkflowVersioningMode` to `WorkerVersioningMode` and its
`VERSIONING_BEHAVIORS` enum to `VERSIONED`, for the following reasons:
- This enum is more about the worker than the workflows, whether or not
workflows are versioned or not depends on whether or not the worker
enables versioning.
- We tried to not use the terms "versioned" and "unversioned" for
workers for a while but the attempt did work and we ended up even
creating special `__unversioned__` value for them. Hence, it'd be very
confusing if they don't see some config such as `WorkerVersioningMode`
that directly specifies the worker as versioned or unversioned. The
previous name, while doing the same thing, twisted the naming and I
believe it'll confuse the users.
- Deployment Versions are not created for the unversioned mode, at least
not so far. This naming and updated comments reflects that better.
- WorkflowVersioningMode is easy to be confused with Workflow
VersioningBehavior.

<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**

---------

Co-authored-by: Carly de Frondeville <carly.defrondeville@temporal.io>