Inherit pinned deployment version from parent workflow (#549)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
Child workflows start in parent's pinned Worker Deployment Version.

<!-- Tell your future self why have you made these changes -->
**Why?**
So user do not have to worry about interface compatibility between
pinned parents and children.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**