Add PauseInfo to the pending activity info (#569)
<!-- Describe what has changed in this PR -->
**What changed?**
Add "PauseInfo" message to the PendincgActivityInfo. Message contain
information about who, why and when paused the activity.
Note: if activity is paused by rule id, users will need to call
"DescribeWorkflowRule" to get the details for that rule.

<!-- Tell your future self why have you made these changes -->
**Why?**
Product request, for audit purpose. Note: we don't add this information
to the history.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
No