Add activity options to the pending activity info (#586)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**
Add activity options to the pending activity info

<!-- Tell your future self why have you made these changes -->
**Why?**
If activity options where changed - there is no "official" way to users
to know what are the current options.
UI should display current options, so we will provide them as a part of
DescribeWorfklow response.
Thus we add activity options to the pending activity info, which will be
a part of DescribeWorfklow response.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
No