Add a new workflow failure cause for oversized grpc messages (#591)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
Add a new workflow failure cause for oversized grpc messages

<!-- Tell your future self why have you made these changes -->
We want to be able to distinguish failures with this cause


<!-- Are there any breaking changes on binary or code level? -->



<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->