inherit pinned version and override across CaN when pinned (#588)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

**What changed?**
Add fields to workflow started event to enable inheriting pinned version
and override across CaN when the effective behavior of the WF is pinned.

This inheritance will not happen if the behavior is anything else
(auto-upgrade, pinned-until-continue-as-new, etc)


**Why?**
To differentiate pinned from trampolining


<!-- Are there any breaking changes on binary or code level? -->
No


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**