Add Stats to DescribeWorkerDeploymentVersion (#603)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**

(1) Deprecated `task_queue_infos` in
`deployment.WorkerDeploymentVersionInfo`.
(2) Added `version_task_queues` to
`DescribeWorkerDeploymentVersionResponse`.

<!-- Tell your future self why have you made these changes -->
**Why?**

We want to report task queue stats for each task queue that is part of a
worker deployment version.

The challenge is that the `taskqueue` package depends on the
`deployment` package. So adding `TaskQueueStats` to
`deployment.WorkerDeploymentVersionInfo` causes a cycle import error.

Weighing our options, we decided to effectively _move_ the task
queue-related data from within the deployment package into the response
message.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**

Not yet; but in subsequent releases the deprecated field
`task_queue_infos` will be removed.

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**

https://github.com/temporalio/temporal/pull/7959 (draft)

---------

Co-authored-by: Spencer Judge <sjudge@hey.com>