Add process key, make heartbeats repeatable field (#608)
<!-- Describe what has changed in this PR -->
**What changed?**

Add "process_key" to the worker host info.
Make "worker_heartbeat" repeatable in "PollNexus" and "RecordHeartbeat"
APIs.

<!-- Tell your future self why have you made these changes -->
**Why?**

We decide to have a single Q per process. Thus we need a way to agree on
task queue name, and for that we need this information to be:
* available as a part of heartbeat
* unique across at least namespace
Process ID is not unique.

Since we send all heartbeats at once - we make them repeated.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**
Not yet.