TaskQueue per-priority stats (#611)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
**What changed?**

Adding a breakdown of stats per-priority levels to the two APIs that
return task queue stats already.

<!-- Tell your future self why have you made these changes -->
**Why?**

To give users insights on a per priority level.

<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**

N/A

<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**

https://github.com/temporalio/temporal/pull/8030