Allow SetCurrent and SetRamping to a version that has never seen pollers, if opt in (#634)
_**READ BEFORE MERGING:** All PRs require approval by both Server AND
SDK teams before merging! This is why the number of required approvals
is "2" and not "1"--two reviewers from the same team is NOT sufficient.
If your PR is not approved by someone in BOTH teams, it may be summarily
reverted._

<!-- Describe what has changed in this PR -->
Allow SetCurrent and SetRamping to a version that has never seen
pollers, if opt in


<!-- Tell your future self why have you made these changes -->
So that users can call SetCurrent or SetRamping as part of their
Deployment pipelines, before the workers have actually been deployed and
started polling. This should only be available to users who opt in and
understand that their tasks won't be picked up until the pollers arrive.

This has been requested by multiple users.


<!-- Are there any breaking changes on binary or code level? -->
**Breaking changes**


<!-- If this breaks the Server, please provide the Server PR to merge
right after this PR was merged. -->
**Server PR**