add dispatch ID parameter to create-release.yaml GH action (#641)
**What changed?**

Added the input param dispatch_id to the create-release.yaml GH action.

**Why?**

This parameter is required by the launchpad release tooling to track GH
actions that are run as part of Temporal WFs.


<https://temporal.io/blog/running-github-actions-temporal-guide#the-dispatch-id-challenge>

**Breaking changes**

None. A default value of "none" is provided. The new parameter can be
ignored when the action is run manually.